refactor(logo): exoprt logo components from index
