+++ components/layout/Footer.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import { Button, LinkNav } from '../ui'
-import TinaWordmark from '../logo/TinaWordmark'
+import { TinaWordmark } from '../logo'
 import TwitterIconSvg from '../../public/svg/twitter-icon.svg'
 import GithubIconSvg from '../../public/svg/github-icon.svg'
 import { EmailForm } from '../forms'
