+++ components/logo/index.ts
@@ -1,0 +1,2 @@
+export * from './TinaIcon'
+export * from './TinaWordmark'
